Split ConstraintsRepository into separate ConstraintsModule installed in ActivityRetainedComponent to address PR feedback
