+++ data/settings/src/main/java/com/google/jetpackcamera/settings/SettingsModule.kt
@@ -18,35 +18,17 @@ package com.google.jetpackcamera.settings
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
-import dagger.hilt.android.components.ActivityRetainedComponent
-import dagger.hilt.android.scopes.ActivityRetainedScoped
+import dagger.hilt.components.SingletonComponent
 
 /**
  * Dagger [Module] for settings data layer.
  */
 @Module
-@InstallIn(ActivityRetainedComponent::class)
+@InstallIn(SingletonComponent::class)
 interface SettingsModule {
 
     @Binds
     fun bindsSettingsRepository(
         localSettingsRepository: LocalSettingsRepository
     ): SettingsRepository
-
-    @Binds
-    @ActivityRetainedScoped
-    fun bindsSettableConstraintsRepository(
-        settableConstraintsRepository: SettableConstraintsRepositoryImpl
-    ): SettableConstraintsRepository
-
-    /**
-     * ConstraintsRepository without setter.
-     *
-     * This is the same instance as the singleton SettableConstraintsRepository, but does not
-     * have the ability to update the constraints.
-     */
-    @Binds
-    fun bindsConstraintsRepository(
-        constraintsRepository: SettableConstraintsRepository
-    ): ConstraintsRepository
 }
